+++ app/test/shared/test_models.dart
@@ -207,7 +207,7 @@ dependencies:
 Iterable<Model> pvModels(PackageVersion pv) sync* {
   yield pv;
   yield _pvPubspec(pv);
-  yield _pvInfo(pv);
+  yield pvToInfo(pv);
 }
 
 PackageVersionPubspec _pvPubspec(PackageVersion pv) {
@@ -219,7 +219,7 @@ PackageVersionPubspec _pvPubspec(PackageVersion pv) {
     ..pubspec = pv.pubspec;
 }
 
-PackageVersionInfo _pvInfo(PackageVersion pv) {
+PackageVersionInfo pvToInfo(PackageVersion pv) {
   return PackageVersionInfo()
     ..parentKey = pv.parentKey.parent
     ..initFromKey(pv.qualifiedVersionKey)
@@ -227,6 +227,30 @@ PackageVersionInfo _pvInfo(PackageVersion pv) {
     ..updated = pv.created
     ..libraries = pv.libraries
     ..libraryCount = pv.libraries.length;
+}
+
+PackageVersionAsset pvToAsset(PackageVersion pv, String assetKind) {
+  PackageVersionAsset convert(FileObject file) {
+    if (file == null) return null;
+    return PackageVersionAsset.init(
+      package: pv.package,
+      version: pv.version,
+      kind: assetKind,
+      versionCreated: pv.created,
+      path: file.filename,
+      textContent: file.text,
+    );
+  }
+
+  if (assetKind == AssetKind.readme) {
+    return convert(pv.readme);
+  } else if (assetKind == AssetKind.changelog) {
+    return convert(pv.changelog);
+  } else if (assetKind == AssetKind.example) {
+    return convert(pv.example);
+  } else {
+    throw ArgumentError('Unhandled asset kind: $assetKind');
+  }
 }
 
 final exampleComPublisher = publisher('example.com');