+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data-models, configuration and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Package page is rendered using `PackageVersionInfo` and `PackageVersionAsset`.
 
 ## `20201020t163909-all`
  * Bumped runtimeVersion to `2020.10.19`.